Modernized code v1

+++ QLepp2DLib/model.h
@@ -101,7 +101,7 @@ public:
     * @param angle p_angle: Provided angle.
     * @return True if method has run without issues.
     */
-    bool detectBadTriangles(double angle);
+    bool detectBadTriangles(float angle);
 
     /**
     * @brief Improves the actual triangulation, modifying the vectors in the Model.
